Port pdq to Windows through x-compilation on Linux
Note that due ton Windows' non-compliance with POSIX, I've had to add my
own getline implementation that is definitely not my own, but is in the
public domain so it should be okay. Thanks Will and Antti.

Also note that in doing the port, the call to random() in
pdqhashtypes.cpp is replaced with a call to rand(), as once again
Windows does not implement the superior random() function. This needs to
be changed in the future.